src/requirements.txt
strands-agents>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
boto3>=1.34.0 